# src/frontend/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ocelescope Backend",
    "version": "0.1"
  },
  "paths": {
...
    "/tasks/plugin": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "returns all tasks of a session",
        "operationId": "getPluginTasks",
        "parameters": [
          {
            "name": "plugin_id",
            "in": "query",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plugin Id"
            }
          },
          {
            "name": "method_name",
            "in": "query",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Method Name"
            }
          },
          {
            "name": "only_running",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Only Running"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PluginTaskSummary"
                  },
                  "title": "Response Getplugintasks"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/plugin/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "returns the task of a given taskId",
        "operationId": "getPluginTask",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginTaskSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plugins/": {
      "get": {
        "tags": [
          "plugins"
        ],
        "summary": "Get Plugins",
        "operationId": "plugins",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PluginApi"
                  },
                  "type": "array",
                  "title": "Response Plugins"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "plugins"
        ],
        "summary": "Upload Plugin",
        "operationId": "uploadPlugin",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadPlugin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plugins/{plugin_id}": {
      "get": {
        "tags": [
          "plugins"
        ],
        "summary": "Get Plugin",
        "operationId": "getPlugin",
        "parameters": [
          {
            "name": "plugin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plugin Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/PluginApi"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Getplugin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "plugins"
        ],
        "summary": "Delete Plugin",
        "operationId": "deletePlugin",
        "parameters": [
          {
            "name": "plugin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plugin Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plugins/{plugin_id}/{method_name}": {
      "get": {
        "tags": [
          "plugins"
        ],
        "summary": "Get Plugin Method",
        "operationId": "getPluginMethod",
        "parameters": [
          {
            "name": "plugin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plugin Id"
            }
          },
          {
            "name": "method_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Method Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/PluginMethod"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Getpluginmethod"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "plugins"
        ],
        "summary": "Run Plugin",
        "operationId": "runPlugin",
        "parameters": [
          {
            "name": "plugin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plugin Id"
            }
          },
          {
            "name": "method_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Method Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_runPlugin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Runplugin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plugins/{plugin_id}/{method_name}/computed/{provider}": {
      "post": {
        "tags": [
          "plugins"
        ],
        "summary": "Get Computed",
        "operationId": "getComputedValues",
        "parameters": [
          {
            "name": "plugin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plugin Id"
            }
          },
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          },
          {
            "name": "method_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Method Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_getComputedValues"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Getcomputedvalues"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/resources/": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Get Resources",
        "operationId": "resources",
        "parameters": [
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResourceApi"
                  },
                  "title": "Response Resources"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "resources"
        ],
        "summary": "Upload Resource",
        "operationId": "uploadResource",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/resources/resource/{resource_id}": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Get Resource",
        "operationId": "resource",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetResourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "resources"
        ],
        "summary": "Delete Resource",
        "operationId": "deleteResource",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "resources"
        ],
        "summary": "Rename Resource",
        "operationId": "renameResource",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          },
          {
            "name": "new_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "New Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_getComputedValues": {
        "properties": {
          "input_ocels": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Input Ocels"
          },
          "input_resources": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Input Resources"
          },
          "input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input"
          }
        },
        "type": "object",
        "required": [
          "input_ocels",
          "input_resources",
          "input"
        ],
        "title": "Body_getComputedValues"
      },
      "Body_runPlugin": {
        "properties": {
          "input_ocels": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Input Ocels"
          },
          "input_resources": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Input Resources"
          },
          "input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "input_ocels",
          "input_resources"
        ],
        "title": "Body_runPlugin"
      },
...
      "Body_uploadPlugin": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_uploadPlugin"
      },
      "Body_uploadResource": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_uploadResource"
      },
...
      "GetResourceResponse": {
        "properties": {
          "resource": {
            "$ref": "#/components/schemas/ResourceApi"
          },
          "visualization": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Graph"
                  },
                  {
                    "$ref": "#/components/schemas/Table"
                  },
                  {
                    "$ref": "#/components/schemas/SVGVis"
                  },
                  {
                    "$ref": "#/components/schemas/DotVis"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "dot": "#/components/schemas/DotVis",
                    "graph": "#/components/schemas/Graph",
                    "svg": "#/components/schemas/SVGVis",
                    "table": "#/components/schemas/Table"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Visualization"
          }
        },
        "type": "object",
        "required": [
          "resource",
          "visualization"
        ],
        "title": "GetResourceResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OCELAnnotation": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "extension": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension"
          }
        },
        "type": "object",
        "required": [
          "label"
        ],
        "title": "OCELAnnotation"
      },
...
      "OCELResult": {
        "properties": {
          "type": {
            "type": "string",
            "const": "ocel",
            "title": "Type",
            "default": "ocel"
          },
          "is_list": {
            "type": "boolean",
            "title": "Is List"
          },
          "annotation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OCELAnnotation"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "is_list",
          "annotation"
        ],
        "title": "OCELResult"
      },
...
      "PluginApi": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "meta": {
            "$ref": "#/components/schemas/PluginMeta"
          },
          "methods": {
            "items": {
              "$ref": "#/components/schemas/PluginMethod"
            },
            "type": "array",
            "title": "Methods"
          }
        },
        "type": "object",
        "required": [
          "id",
          "meta",
          "methods"
        ],
        "title": "PluginApi"
      },
      "PluginMeta": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name",
          "version",
          "label",
          "description"
        ],
        "title": "PluginMeta"
      },
      "PluginMethod": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "input_ocels": {
            "additionalProperties": {
              "$ref": "#/components/schemas/OCELAnnotation"
            },
            "type": "object",
            "title": "Input Ocels"
          },
          "input_resources": {
            "additionalProperties": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "$ref": "#/components/schemas/ResourceAnnotation"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "object",
            "title": "Input Resources"
          },
          "results": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/OCELResult"
                },
                {
                  "$ref": "#/components/schemas/ResourceResult"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "ocel": "#/components/schemas/OCELResult",
                  "resource": "#/components/schemas/ResourceResult"
                }
              }
            },
            "type": "array",
            "title": "Results"
          },
          "input_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Schema",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "input_schema"
        ],
        "title": "PluginMethod"
      },
      "PluginOutput": {
        "properties": {
          "ocel_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ocel Ids",
            "default": []
          },
          "resource_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Resource Ids",
            "default": []
          }
        },
        "type": "object",
        "title": "PluginOutput"
      },
...
      "PluginTaskSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "state": {
            "$ref": "#/components/schemas/TaskState"
          },
          "plugin_id": {
            "type": "string",
            "title": "Plugin Id"
          },
          "method_name": {
            "type": "string",
            "title": "Method Name"
          },
          "output": {
            "$ref": "#/components/schemas/PluginOutput"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state",
          "plugin_id",
          "method_name",
          "output"
        ],
        "title": "PluginTaskSummary"
      },
...
      "ResourceAnnotation": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "label"
        ],
        "title": "ResourceAnnotation"
      },
      "ResourceApi": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PluginSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "type",
          "name",
          "source",
          "id"
        ],
        "title": "ResourceApi"
      },
...
      "ResourceResult": {
        "properties": {
          "type": {
            "type": "string",
            "const": "resource",
            "title": "Type",
            "default": "resource"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "is_list": {
            "type": "boolean",
            "title": "Is List"
          },
          "annotation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceAnnotation"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "resource_type",
          "is_list",
          "annotation"
        ],
        "title": "ResourceResult"
      },
...
      "TaskState": {
        "type": "string",
        "enum": [
          "PENDING",
          "STARTED",
          "SUCCESS",
          "FAILURE",
          "SKIPPED",
          "CANCELLED"
        ],
        "title": "TaskState"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}